Setup
Fixed TypeScript React
Fixed Tailwind
Fixed colors used in design as CSS classes
