+++ src/App.tsx
@@ -1,0 +1,8 @@
+function App() {
+  return (
+    <div className="bg-blue-300">
+    </div>
+  );
+}
+
+export default App;
