+++ tailwind.config.js
@@ -1,0 +1,10 @@
+/** @type {import('tailwindcss').Config} */ 
+module.exports = {
+  content: [
+    "./src/**/*.{js,jsx,ts,tsx}",
+  ],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+}
